feat: Get email sending api working with name
